+++ src/SignalR/clients/ts/signalr/tests/HubConnection.test.ts
@@ -13,6 +13,7 @@ import { Subject } from "../src/Subject";
 import { TextMessageFormat } from "../src/TextMessageFormat";
 
 import { VerifyLogger } from "./Common";
+import { TestConnection } from "./TestConnection";
 import { delayUntil, PromiseSource, registerUnhandledRejectionHandler } from "./Utils";
 
 function createHubConnection(connection: IConnection, logger?: ILogger | null, protocol?: IHubProtocol | null) {
@@ -1236,77 +1237,6 @@ describe("HubConnection", () => {
 async function pingAndWait(connection: TestConnection): Promise<void> {
     await connection.receive({ type: MessageType.Ping });
     await delayUntil(50);
-}
-
-class TestConnection implements IConnection {
-    public readonly features: any = {};
-    public connectionId?: string;
-    public onreceive: ((data: string | ArrayBuffer) => void) | null;
-    public onclose: ((error?: Error) => void) | null;
-    public sentData: any[];
-    public lastInvocationId: string | null;
-
-    private autoHandshake: boolean | null;
-
-    constructor(autoHandshake: boolean = true) {
-        this.onreceive = null;
-        this.onclose = null;
-        this.sentData = [];
-        this.lastInvocationId = null;
-        this.autoHandshake = autoHandshake;
-    }
-
-    public start(): Promise<void> {
-        return Promise.resolve();
-    }
-
-    public send(data: any): Promise<void> {
-        const invocation = TextMessageFormat.parse(data)[0];
-        const parsedInvocation = JSON.parse(invocation);
-        const invocationId = parsedInvocation.invocationId;
-        if (parsedInvocation.protocol && parsedInvocation.version && this.autoHandshake) {
-            this.receiveHandshakeResponse();
-        }
-        if (invocationId) {
-            this.lastInvocationId = invocationId;
-        }
-        if (this.sentData) {
-            this.sentData.push(invocation);
-        } else {
-            this.sentData = [invocation];
-        }
-        return Promise.resolve();
-    }
-
-    public stop(error?: Error): Promise<void> {
-        if (this.onclose) {
-            this.onclose(error);
-        }
-        return Promise.resolve();
-    }
-
-    public receiveHandshakeResponse(error?: string): void {
-        this.receive({ error });
-    }
-
-    public receive(data: any): void {
-        const payload = JSON.stringify(data);
-        this.invokeOnReceive(TextMessageFormat.write(payload));
-    }
-
-    public receiveText(data: string) {
-        this.invokeOnReceive(data);
-    }
-
-    public receiveBinary(data: ArrayBuffer) {
-        this.invokeOnReceive(data);
-    }
-
-    private invokeOnReceive(data: string | ArrayBuffer) {
-        if (this.onreceive) {
-            this.onreceive(data);
-        }
-    }
 }
 
 class TestProtocol implements IHubProtocol {
